refactor(module:tree-select): refactor tree-select component
