+++ components/tree-select/nz-tree-select.component.ts
@@ -1,16 +1,11 @@
 import { BACKSPACE } from '@angular/cdk/keycodes';
 import {
-  ConnectionPositionPair,
-  FlexibleConnectedPositionStrategy,
-  Overlay,
-  OverlayConfig,
-  OverlayRef,
-  PositionStrategy
+  CdkConnectedOverlay,
+  CdkOverlayOrigin,
+  ConnectedOverlayPositionChange
 } from '@angular/cdk/overlay';
-import { TemplatePortal } from '@angular/cdk/portal';
 import {
   forwardRef,
-  AfterViewInit,
   ChangeDetectorRef,
   Component,
   ElementRef,
@@ -23,7 +18,6 @@ import {
   Output,
   Renderer2,
   SimpleChanges,
-  TemplateRef,
   ViewChild,
   ViewContainerRef
 } from '@angular/core';
@@ -40,14 +34,16 @@ import { selectDropDownAnimation } from '../core/animation/select-dropdown-anima
 import { zoomMotion } from '../core/animation/zoom';
 import { InputBoolean } from '../core/util/convert';
 import { NzFormatEmitEvent } from '../tree/interface';
-import { NzTreeNode } from '../tree/nz-tree-node';
+import { NzTreeNode, NzTreeNodeOptions } from '../tree/nz-tree-node';
 import { NzTreeComponent } from '../tree/nz-tree.component';
+import { NzTreeSelectService } from './nz-tree-select.service';
 
 @Component({
   selector   : 'nz-tree-select',
   animations : [ selectDropDownAnimation, zoomMotion ],
   templateUrl: './nz-tree-select.component.html',
   providers  : [
+    NzTreeSelectService,
     {
       provide    : NG_VALUE_ACCESSOR,
       useExisting: forwardRef(() => NzTreeSelectComponent),
@@ -74,7 +70,7 @@ import { NzTreeComponent } from '../tree/nz-tree.component';
     }
   ` ]
 })
-export class NzTreeSelectComponent implements ControlValueAccessor, OnInit, AfterViewInit, OnDestroy, OnChanges {
+export class NzTreeSelectComponent implements ControlValueAccessor, OnInit, OnDestroy, OnChanges {
 
   @Input() @InputBoolean() nzAllowClear = true;
   @Input() @InputBoolean() nzShowExpand = true;
@@ -87,12 +83,30 @@ export class NzTreeSelectComponent implements ControlValueAccessor, OnInit, Afte
   @Input() @InputBoolean() nzMultiple = false;
   @Input() @InputBoolean() nzDefaultExpandAll = false;
   @Input() nzNotFoundContent: string;
-  @Input() nzNodes: NzTreeNode[] = [];
+  @Input() nzNodes: Array<NzTreeNode | NzTreeNodeOptions> = [];
   @Input() nzOpen = false;
   @Input() nzSize = 'default';
   @Input() nzPlaceHolder = '';
   @Input() nzDropdownStyle: { [ key: string ]: string; };
-  @Input() nzDefaultExpandedKeys: string[] = [];
+
+  @Input()
+  get nzDefaultExpandedKeys(): string[] {
+    return this.expandedKeys;
+  }
+
+  set nzDefaultExpandedKeys(value: string[]) {
+    this.expandedKeys = value;
+  }
+
+  @Input()
+  get nzExpandedKeys(): string[] {
+    return this.expandedKeys;
+  }
+
+  set nzExpandedKeys(value: string[]) {
+    this.expandedKeys = value;
+  }
+
   @Input() nzDisplayWith: (node: NzTreeNode) => string = (node: NzTreeNode) => node.title;
   @Output() readonly nzOpenChange = new EventEmitter<boolean>();
   @Output() readonly nzCleared = new EventEmitter<void>();
@@ -102,23 +116,20 @@ export class NzTreeSelectComponent implements ControlValueAccessor, OnInit, Afte
   @Output() readonly nzTreeCheckBoxChange = new EventEmitter<NzFormatEmitEvent>();
 
   @ViewChild('inputElement') inputElement: ElementRef;
-  @ViewChild('treeSelect') treeSelect: ElementRef;
-  @ViewChild('dropdownTemplate', { read: TemplateRef }) dropdownTemplate;
   @ViewChild('treeRef') treeRef: NzTreeComponent;
+  @ViewChild(CdkOverlayOrigin) cdkOverlayOrigin: CdkOverlayOrigin;
+  @ViewChild(CdkConnectedOverlay) cdkConnectedOverlay: CdkConnectedOverlay;
 
+  triggerWidth: number;
   isComposing = false;
   isDestroy = true;
   isNotFound = false;
   inputValue = '';
-  dropDownClassMap: { [ className: string ]: boolean };
   dropDownPosition: 'top' | 'center' | 'bottom' = 'bottom';
-  overlayRef: OverlayRef;
-  portal: TemplatePortal<{}>;
-  positionStrategy: FlexibleConnectedPositionStrategy;
-  overlayBackdropClickSubscription: Subscription;
   selectionChangeSubscription: Subscription;
   selectedNodes: NzTreeNode[] = [];
   value: string[] = [];
+  expandedKeys: string[] = [];
 
   onChange: (value: string[] | string) => void;
   onTouched: () => void = () => null;
@@ -159,7 +170,7 @@ export class NzTreeSelectComponent implements ControlValueAccessor, OnInit, Afte
   constructor(
     private renderer: Renderer2,
     private cdr: ChangeDetectorRef,
-    private overlay: Overlay,
+    private nzTreeService: NzTreeSelectService,
     private viewContainerRef: ViewContainerRef,
     elementRef: ElementRef) {
     renderer.addClass(elementRef.nativeElement, 'ant-select');
@@ -169,20 +180,14 @@ export class NzTreeSelectComponent implements ControlValueAccessor, OnInit, Afte
     this.isDestroy = false;
     this.selectionChangeSubscription = this.subscribeSelectionChange();
     Promise.resolve().then(() => {
-      this.updateDropDownClassMap();
-      this.updateCdkConnectedOverlayStatus();
+      this.updateTreeClass();
     });
   }
 
   ngOnDestroy(): void {
     this.isDestroy = true;
-    this.detachOverlay();
+    this.closeDropDown();
     this.selectionChangeSubscription.unsubscribe();
-    this.overlayBackdropClickSubscription.unsubscribe();
-  }
-
-  ngAfterViewInit(): void {
-    this.attachOverlay();
   }
 
   setDisabledState(isDisabled: boolean): void {
@@ -192,7 +197,7 @@ export class NzTreeSelectComponent implements ControlValueAccessor, OnInit, Afte
 
   ngOnChanges(changes: SimpleChanges): void {
     if (changes.hasOwnProperty('nzNodes')) {
-      setTimeout(() => this.updateSelectedNodes(), 0);
+      setTimeout(() => this.updateSelectedNodes(true), 0);
     }
   }
 
@@ -203,7 +208,7 @@ export class NzTreeSelectComponent implements ControlValueAccessor, OnInit, Afte
       } else {
         this.value = [ (value as string) ];
       }
-      this.updateSelectedNodes();
+      this.updateSelectedNodes(true);
     } else {
       this.value = [];
       this.selectedNodes.forEach(node => {
@@ -239,9 +244,8 @@ export class NzTreeSelectComponent implements ControlValueAccessor, OnInit, Afte
       this.nzOpen = true;
       this.nzOpenChange.emit(this.nzOpen);
       this.updateCdkConnectedOverlayStatus();
+      this.updateTreeClass();
       this.updatePosition();
-      this.updateDropDownClassMap();
-      this.updateStackingOrder();
     }
   }
 
@@ -249,7 +253,6 @@ export class NzTreeSelectComponent implements ControlValueAccessor, OnInit, Afte
     this.onTouched();
     this.nzOpen = false;
     this.nzOpenChange.emit(this.nzOpen);
-    this.updateCdkConnectedOverlayStatus();
     this.cdr.markForCheck();
   }
 
@@ -268,31 +271,27 @@ export class NzTreeSelectComponent implements ControlValueAccessor, OnInit, Afte
     }
   }
 
+  onExpandedKeysChange(value: NzFormatEmitEvent): void {
+    this.nzExpandChange.emit(value);
+    this.expandedKeys = [ ...value.keys ];
+  }
+
   setInputValue(value: string): void {
     this.inputValue = value;
     this.updateInputWidth();
     this.updatePosition();
   }
 
-  detachOverlay(): void {
-    if (this.overlayRef && this.overlayRef.hasAttached()) {
-      this.overlayRef.detach();
-      this.overlayBackdropClickSubscription.unsubscribe();
-      this.onTouched();
-      this.nzOpen = false;
-      this.nzOpenChange.emit(this.nzOpen);
-    }
-  }
-
   removeSelected(node: NzTreeNode, emit: boolean = true, event?: MouseEvent): void {
     node.isSelected = false;
     node.isChecked = false;
     if (this.nzCheckable) {
-      this.treeRef.nzTreeService.conduct(node);
-      this.treeRef.nzTreeService.setCheckedNodeList(node);
+      this.nzTreeService.conduct(node);
+      this.nzTreeService.setCheckedNodeList(node);
     } else {
-      this.treeRef.nzTreeService.setSelectedNodeList(node, this.nzMultiple);
+      this.nzTreeService.setSelectedNodeList(node, this.nzMultiple);
     }
+
     if (emit) {
       this.nzRemoved.emit(node);
     }
@@ -311,52 +310,6 @@ export class NzTreeSelectComponent implements ControlValueAccessor, OnInit, Afte
     });
   }
 
-  updateStackingOrder(): void {
-    if (this.renderer.nextSibling(this.overlayRef.hostElement)) {
-      const parentNode = this.renderer.parentNode(this.overlayRef.hostElement);
-      this.renderer.appendChild(parentNode, this.overlayRef.backdropElement);
-      this.renderer.appendChild(parentNode, this.overlayRef.hostElement);
-    }
-  }
-
-  attachOverlay(): void {
-    this.portal = new TemplatePortal(this.dropdownTemplate, this.viewContainerRef);
-    this.overlayRef = this.overlay.create(this.getOverlayConfig());
-    this.overlayRef.attach(this.portal);
-    this.cdr.detectChanges();
-    this.overlayBackdropClickSubscription = this.subscribeOverlayBackdropClick();
-  }
-
-  getOverlayConfig(): OverlayConfig {
-    const overlayWidth = this.treeSelect.nativeElement.getBoundingClientRect().width;
-    return new OverlayConfig({
-      positionStrategy                                          : this.getOverlayPosition(),
-      scrollStrategy                                            : this.overlay.scrollStrategies.reposition(),
-      [ this.nzDropdownMatchSelectWidth ? 'width' : 'minWidth' ]: overlayWidth,
-      hasBackdrop                                               : true
-    });
-  }
-
-  getOverlayPosition(): PositionStrategy {
-    const positions = [
-      new ConnectionPositionPair({ originX: 'start', originY: 'bottom' }, { overlayX: 'start', overlayY: 'top' }),
-      new ConnectionPositionPair({ originX: 'start', originY: 'top' }, { overlayX: 'start', overlayY: 'bottom' })
-    ];
-    this.positionStrategy = this.overlay.position()
-    .flexibleConnectedTo(this.treeSelect)
-    .withPositions(positions)
-    .withFlexibleDimensions(false)
-    .withPush(false);
-    return this.positionStrategy;
-  }
-
-  subscribeOverlayBackdropClick(): Subscription {
-    return this.overlayRef.backdropClick()
-    .subscribe(() => {
-      this.closeDropDown();
-    });
-  }
-
   subscribeSelectionChange(): Subscription {
     return merge(
       this.nzTreeClick.pipe(
@@ -364,8 +317,8 @@ export class NzTreeSelectComponent implements ControlValueAccessor, OnInit, Afte
           const node = event.node;
           if (this.nzCheckable && !node.isDisabled && !node.isDisableCheckbox) {
             node.isChecked = !node.isChecked;
-            this.treeRef.nzTreeService.conduct(node);
-            this.treeRef.nzTreeService.setCheckedNodeList(node);
+            this.nzTreeService.conduct(node);
+            this.nzTreeService.setCheckedNodeList(node);
           }
           if (this.nzCheckable) {
             node.isSelected = false;
@@ -389,6 +342,7 @@ export class NzTreeSelectComponent implements ControlValueAccessor, OnInit, Afte
       if (this.isMultiple) {
         this.onChange(value);
         this.focusOnInput();
+        this.updatePosition();
       } else {
         this.closeDropDown();
         this.onChange(value.length ? value[ 0 ] : null);
@@ -397,14 +351,29 @@ export class NzTreeSelectComponent implements ControlValueAccessor, OnInit, Afte
     });
   }
 
-  updateSelectedNodes(): void {
-    if (this.treeRef) {
-      this.selectedNodes = [ ...(this.nzCheckable ? this.treeRef.getCheckedNodeList() : this.treeRef.getSelectedNodeList()) ];
+  updateSelectedNodes(init: boolean = false): void {
+    if (init) {
+      const nodes = this.nzNodes.map(item => new NzTreeNode(item));
+      this.nzTreeService.initTree(nodes);
+      if (this.nzCheckable) {
+        this.nzTreeService.calcCheckedKeys(this.value, nodes);
+      } else {
+        this.nzTreeService.calcSelectedKeys(this.value, nodes);
+      }
     }
+    this.selectedNodes = [ ...(this.nzCheckable ? this.nzTreeService.getCheckedNodeList() : this.nzTreeService.getSelectedNodeList()) ];
   }
 
   updatePosition(): void {
-    this.overlayRef.updatePosition();
+    setTimeout(() => {
+      if (this.cdkConnectedOverlay && this.cdkConnectedOverlay.overlayRef) {
+        this.cdkConnectedOverlay.overlayRef.updatePosition();
+      }
+    });
+  }
+
+  onPositionChange(position: ConnectedOverlayPositionChange): void {
+    this.dropDownPosition = position.connectionPair.originY;
   }
 
   updateInputWidth(): void {
@@ -417,12 +386,13 @@ export class NzTreeSelectComponent implements ControlValueAccessor, OnInit, Afte
     }
   }
 
-  onClearSelection(): void {
+  onClearSelection($event: MouseEvent): void {
+    $event.stopPropagation();
+    $event.preventDefault();
     this.selectedNodes.forEach(node => {
       this.removeSelected(node, false);
     });
     this.nzCleared.emit();
-    this.closeDropDown();
   }
 
   setSearchValues($event: NzFormatEmitEvent): void {
@@ -433,32 +403,18 @@ export class NzTreeSelectComponent implements ControlValueAccessor, OnInit, Afte
     });
   }
 
-  updateDropDownClassMap(): void {
+  updateTreeClass(): void {
     if (this.treeRef && !this.treeRef.nzTreeClass[ 'ant-select-tree' ]) {
       this.treeRef.nzTreeClass = { ...this.treeRef.nzTreeClass, [ 'ant-select-tree' ]: true };
     }
-    this.dropDownClassMap = {
-      [ 'ant-select-dropdown' ]                     : true,
-      [ 'ant-select-tree-dropdown' ]                : true,
-      [ `ant-select-dropdown--single` ]             : !this.nzMultiple,
-      [ `ant-select-dropdown--multiple` ]           : this.nzMultiple,
-      [ `ant-select-dropdown-placement-bottomLeft` ]: this.dropDownPosition === 'bottom',
-      [ `ant-select-dropdown-placement-topLeft` ]   : this.dropDownPosition === 'top'
-    };
   }
 
   updateCdkConnectedOverlayStatus(): void {
-    const overlayWidth = this.treeSelect.nativeElement.getBoundingClientRect().width;
-    if (this.nzDropdownMatchSelectWidth) {
-      this.overlayRef.updateSize({ width: overlayWidth });
-    } else {
-      this.overlayRef.updateSize({ minWidth: overlayWidth });
-    }
+    this.triggerWidth = this.cdkOverlayOrigin.elementRef.nativeElement.getBoundingClientRect().width;
+  }
 
-    if (this.nzOpen) {
-      this.renderer.removeStyle(this.overlayRef.backdropElement, 'display');
-    } else {
-      this.renderer.setStyle(this.overlayRef.backdropElement, 'display', 'none');
-    }
+  // tslint:disable-next-line:no-any
+  trackValue(index: number, option: NzTreeNode): any {
+    return option.key;
   }
 }
