+++ components/tree-select/nz-tree-select.service.ts
@@ -1,0 +1,7 @@
+import { Injectable } from '@angular/core';
+import { NzTreeBaseService } from '../tree/nz-tree-base.service';
+
+@Injectable()
+export class NzTreeSelectService extends NzTreeBaseService {
+
+}
