+++ components/tree/nz-tree-node.component.html
@@ -75,8 +75,11 @@
 
   <ul
     role="group"
+    class="ant-tree-child-tree"
+    [class.ant-tree-child-tree-open]="nzTreeNode.isExpanded"
     [attr.data-expanded]="nzTreeNode.isExpanded"
     [ngClass]="nzNodeChildrenClass"
+    [@.disabled]="nzNoopAnimations"
     [@nodeState]="nzTreeNode.isExpanded ? 'active' : 'inactive'">
     <nz-tree-node
       *ngFor="let node of nzTreeNode.getChildren()"
