+++ components/tree/nz-tree.component.html
@@ -18,6 +18,7 @@
     [nzExpandAll]="nzExpandAll"
     [nzDefaultExpandAll]="nzDefaultExpandAll"
     [nzTreeTemplate]="nzTreeTemplate"
+    [nzNoopAnimations]="nzNoopAnimations"
     (clickNode)="nzClick.emit($event)"
     (dblClick)="nzDblClick.emit($event)"
     (contextMenu)="nzContextMenu.emit($event)"
