+++ components/tree/public-api.ts
@@ -3,5 +3,5 @@ export * from './nz-tree.module';
 export * from './nz-tree.component';
 export * from './nz-tree-node.component';
 export * from './nz-tree-node';
-export * from './nz-tree.service';
+export * from './nz-tree-base.service';
 export * from './nz-tree-util';
